fix: fix based on review comments
